Move list folder to ui folder
