+++ packages/frontend/src/components/ui/lists/ListColumnsHeader.scss
@@ -1,5 +1,5 @@
-@use '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
+@use '../../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
 
 .ListColumnsHeader {
   @include mixins.defListTitles;
