+++ packages/frontend/src/components/ui/lists/SimpleList.scss
@@ -1,5 +1,5 @@
-@use '../../styles/mixins.scss';
-@import '../../styles/variables.scss';
+@use '../../../styles/mixins.scss';
+@import '../../../styles/variables.scss';
 
 .SimpleList {
   &__ColumnTitles {
